better support for content-types
make sure we return a correct content type for the file type, currently
supported

- html
- json
- txt

also json is extended to support comments in a js style syntax so
documentation of resposes can be handles inline. The comments will be
stripped out before serving the json, so the served json will always be
valid.

added support for content type for javascript

added documentation for content-type handling