naming files id is wrong should be any
initially the plan obviously was to use id as the special name, but now
canned is using any as the keyword. So naming the files in example
```id.get.json``` is just wrong, to make them available via ```curl
localhost:3000/comment/1``` they need to be named ```any.get.json``` and
accordingly ```any.put.json```.

This causes confusion as seen in #20