make sure we support multiple files in matches
This is the first step to support windows since this will make moving
the query to the file possible.

You can now create multiple files with a number attached to the end of
the name. Currently the first file, according to file listing, will be
picked, this is WIP so and it should be a random pick.